+++ choreo-extension-native/spotbugs-exclude.xml
@@ -18,26 +18,26 @@
 
 <FindBugsFilter>
     <Match>
-        <Package name="org.ballerinalang.observe.trace.extension.choreo.gen"/>
+        <Package name="io.ballerina.observe.choreo.gen"/>
     </Match>
     <Match>
-        <Class name="org.ballerinalang.observe.trace.extension.choreo.client.ChoreoClientHolder" />
+        <Class name="io.ballerina.observe.choreo.client.ChoreoClientHolder" />
         <Bug pattern="RV_RETURN_VALUE_IGNORED_BAD_PRACTICE,NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE"/>
     </Match>
     <Match>
-        <Class name="org.ballerinalang.observe.trace.extension.choreo.client.secret.AnonymousAppSecretHandler" />
+        <Class name="io.ballerina.observe.choreo.client.secret.AnonymousAppSecretHandler" />
         <Bug pattern="RV_RETURN_VALUE_IGNORED_BAD_PRACTICE,NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE"/>
     </Match>
     <Match>
-        <Class name="org.ballerinalang.observe.trace.extension.choreo.client.secret.AnonymousAppSecretHandler"/>
+        <Class name="io.ballerina.observe.choreo.client.secret.AnonymousAppSecretHandler"/>
     </Match>
     <Match>
-        <Class name="org.ballerinalang.observe.trace.extension.choreo.ChoreoJaegerReporter"/>
+        <Class name="io.ballerina.observe.choreo.ChoreoJaegerReporter"/>
     </Match>
     <Match>
         <Package name="org.ballerinalang.observe.trace.extension.choreo"/>
     </Match>
     <Match>
-        <Class name="org.ballerinalang.observe.trace.extension.choreo.client.MetadataReader"/>
+        <Class name="io.ballerina.observe.choreo.client.MetadataReader"/>
     </Match>
 </FindBugsFilter>
