+++ choreo-extension-native/src/main/java/io/ballerina/observe/choreo/Constants.java
@@ -1,0 +1,48 @@
+
+/*
+ * Copyright (c) 2020, WSO2 Inc. (http://wso2.com) All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.ballerina.observe.choreo;
+
+/**
+ * This is the constants class that defines all the constants
+ * that are used by the {@link ChoreoTracerProvider}.
+ *
+ * @since 2.0.0
+ */
+public class Constants {
+
+    private Constants() {}
+
+    public static final String CHOREO_EXTENSION_NAME = "choreo";
+
+    public static final String EXTENSION_PROPERTIES_FILE = "choreo-extension.properties";
+    public static final String PACKAGE_VERSION_PROPERTY_KEY = "moduleVersion";
+    public static final String PACKAGE_ORG = "ballerinax";
+    public static final String PACKAGE_NAME = "choreo";
+
+    public static final String REPORTER_HOST_NAME_CONFIG = "reporter.hostname";
+    public static final String DEFAULT_REPORTER_HOSTNAME = "periscope.choreo.dev";
+
+    public static final String REPORTER_PORT_CONFIG = "reporter.port";
+    public static final int DEFAULT_REPORTER_PORT = 443;
+
+    public static final String REPORTER_USE_SSL_CONFIG = "reporter.useSSL";
+    public static final boolean DEFAULT_REPORTER_USE_SSL = true;
+
+    public static final String APPLICATION_ID_CONFIG = "application.secret";
+    public static final String EMPTY_APPLICATION_SECRET = "";
+}
