+++ choreo-extension-native/src/main/java/io/ballerina/observe/choreo/client/error/ChoreoClientException.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerina.observe.choreo.client.error;
+
+/**
+ * Choreo Client specific exception class.
+ *
+ * @since 2.0.0
+ */
+public class ChoreoClientException extends Exception {
+
+    private final ChoreoError code;
+
+    /**
+     * Constructs the exception with both a status and trailers.
+     *
+     * @param error Error status
+     *
+     */
+    public ChoreoClientException(ChoreoError error) {
+        super(error.formatThrowableMessage(error), error.getCause());
+        this.code = error;
+    }
+
+    /**
+     * Returns the error code as a {@link ChoreoError} object.
+     *
+     * @return Error status
+     */
+    public final ChoreoError getCode() {
+        return code;
+    }
+}
