--- choreo-extension-native/src/main/java/org/ballerinalang/observe/trace/extension/choreo/BallerinaPackageResourceReader.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2020, WSO2 Inc. (http://wso2.com) All Rights Reserved.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.ballerinalang.observe.trace.extension.choreo;
-
-import java.io.InputStream;
-
-/**
- * Handle resource files in Ballerina program packages.
- *
- * @since 2.0.0
- */
-public class BallerinaPackageResourceReader {
-
-    /**
-     * Get resource as stream from Ballerina program package.
-     *
-     * @param filePath resource pat
-     * @return input stream
-     */
-    public static InputStream getResourceAsStream(String filePath) {
-        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-        return classLoader.getResourceAsStream(filePath);
-    }
-}
