--- choreo-extension-native/src/main/java/org/ballerinalang/observe/trace/extension/choreo/ChoreoTracerProvider.java
@@ -1,58 +1,0 @@
-/*
- * Copyright (c) 2020, WSO2 Inc. (http://wso2.com) All Rights Reserved.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.ballerinalang.observe.trace.extension.choreo;
-
-import io.ballerina.runtime.api.creators.ErrorCreator;
-import io.ballerina.runtime.api.utils.StringUtils;
-import io.ballerina.runtime.observability.tracer.spi.TracerProvider;
-import io.jaegertracing.internal.JaegerTracer;
-import io.jaegertracing.internal.samplers.RateLimitingSampler;
-import io.jaegertracing.spi.Reporter;
-import io.opentracing.Tracer;
-import org.ballerinalang.observe.trace.extension.choreo.client.ChoreoClient;
-import org.ballerinalang.observe.trace.extension.choreo.client.ChoreoClientHolder;
-
-import java.util.Objects;
-
-/**
- * This is the open tracing extension class for {@link TracerProvider}.
- *
- * @since 2.0.0
- */
-public class ChoreoTracerProvider implements TracerProvider {
-    private static volatile Reporter reporterInstance;
-
-    @Override
-    public Tracer getTracer(String serviceName) {
-        ChoreoClient choreoClient = ChoreoClientHolder.getChoreoClient();
-        if (Objects.isNull(choreoClient)) {
-            throw ErrorCreator.createError(StringUtils.fromString(
-                    "Choreo client is not initialized. Please check Ballerina configurations."));
-        }
-
-        return new JaegerTracer.Builder(serviceName)
-                .withSampler(new RateLimitingSampler(2))
-                .withReporter(getReporter())
-                .build();
-    }
-
-    public static Reporter getReporter() {
-        if (reporterInstance == null) {
-            reporterInstance = new ChoreoJaegerReporter();
-        }
-        return reporterInstance;
-    }
-}
