+++ choreo-extension-native/src/main/proto/telemetry.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 package v0_1_2;
 
-option java_package = "org.ballerinalang.observe.trace.extension.choreo.gen";
+option java_package = "io.ballerina.observe.choreo.gen";
 
 import "google/protobuf/empty.proto";
