+++ choreo-extension-native/src/main/resources/META-INF/services/io.ballerina.runtime.observability.metrics.spi.MetricReporterFactory
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-org.ballerinalang.observe.trace.extension.choreo.ChoreoMetricReporterFactory
+io.ballerina.observe.choreo.ChoreoMetricReporterFactory
