+++ choreo-extension-native/src/main/resources/META-INF/services/io.ballerina.runtime.observability.tracer.spi.TracerProviderFactory
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-org.ballerinalang.observe.trace.extension.choreo.ChoreoTracerProviderFactory
+io.ballerina.observe.choreo.ChoreoTracerProviderFactory
